refactor(integrations): split monolithic module into separate discord and twitch modules
- Create separate app-modules/integration-discord and app-modules/integration-twitch modules
- Move Discord OAuth implementation to dedicated module with proper namespace
- Move Twitch OAuth and subscribers implementation to dedicated module
- Update IdentityProvider enum imports to reference new module namespaces
- Remove monolithic integrations module in favor of modular structure
- Update composer.json to require individual integration modules
- Add service providers for each integration module with appropriate bindings
