+++ app-modules/integration-discord/composer.json
@@ -1,0 +1,24 @@
+{
+    "name": "he4rt/integration-discord",
+    "description": "",
+    "type": "library",
+    "version": "1.0",
+    "license": "proprietary",
+    "require": {},
+    "autoload": {
+        "psr-4": {
+            "He4rt\\IntegrationDiscord\\": "src/",
+            "He4rt\\IntegrationDiscord\\Tests\\": "tests/",
+            "He4rt\\IntegrationDiscord\\Database\\Factories\\": "database/factories/",
+            "He4rt\\IntegrationDiscord\\Database\\Seeders\\": "database/seeders/"
+        }
+    },
+    "minimum-stability": "stable",
+    "extra": {
+        "laravel": {
+            "providers": [
+                "He4rt\\IntegrationDiscord\\Providers\\IntegrationDiscordServiceProvider"
+            ]
+        }
+    }
+}
