+++ app-modules/integration-discord/src/Providers/IntegrationDiscordServiceProvider.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace He4rt\IntegrationDiscord\Providers;
+
+use Illuminate\Support\ServiceProvider;
+
+class IntegrationDiscordServiceProvider extends ServiceProvider
+{
+    public function register(): void {}
+
+    public function boot(): void {}
+}
