+++ app-modules/integration-twitch/composer.json
@@ -1,0 +1,24 @@
+{
+    "name": "he4rt/integration-twitch",
+    "description": "",
+    "type": "library",
+    "version": "1.0",
+    "license": "proprietary",
+    "require": {},
+    "autoload": {
+        "psr-4": {
+            "He4rt\\IntegrationTwitch\\": "src/",
+            "He4rt\\IntegrationTwitch\\Tests\\": "tests/",
+            "He4rt\\IntegrationTwitch\\Database\\Factories\\": "database/factories/",
+            "He4rt\\IntegrationTwitch\\Database\\Seeders\\": "database/seeders/"
+        }
+    },
+    "minimum-stability": "stable",
+    "extra": {
+        "laravel": {
+            "providers": [
+                "He4rt\\IntegrationTwitch\\Providers\\IntegrationTwitchServiceProvider"
+            ]
+        }
+    }
+}
