+++ app-modules/integration-twitch/src/OAuth/Contracts/TwitchOAuthService.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace He4rt\IntegrationTwitch\OAuth\Contracts;
+
+use App\Contracts\OAuthClientContract;
+
+interface TwitchOAuthService extends OAuthClientContract {}
