+++ app-modules/integration-twitch/src/Subscriber/Contracts/TwitchSubscribersService.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace He4rt\IntegrationTwitch\Subscriber\Contracts;
+
+use He4rt\Identity\Auth\DTOs\OAuthAccessDTO;
+
+interface TwitchSubscribersService
+{
+    public function getSubscriptionState(OAuthAccessDTO $dto, string $twitchId, string $channelId);
+}
