--- app-modules/integrations/composer.json
@@ -1,21 +1,0 @@
-{
-    "name": "he4rt/integrations",
-    "description": "",
-    "type": "library",
-    "version": "1.0",
-    "license": "proprietary",
-    "require": {},
-    "autoload": {
-        "psr-4": {
-            "He4rt\\Integrations\\": "src/"
-        }
-    },
-    "minimum-stability": "stable",
-    "extra": {
-        "laravel": {
-            "providers": [
-                "He4rt\\Integrations\\Providers\\IntegrationsServiceProvider"
-            ]
-        }
-    }
-}
