--- app-modules/integrations/phpstan.ignore.neon
@@ -1,5 +1,0 @@
-parameters:
-    ignoreErrors:
-    excludePaths:
-        analyse:
-            - src/Twitch/OAuth/Contracts/TwitchOAuthService.php
