--- app-modules/integrations/src/Common/Client/TwitchBaseClient.php
@@ -1,27 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace He4rt\Integrations\Common\Client;
-
-use GuzzleHttp\Client;
-use He4rt\Integrations\Common\Contracts\TwitchService;
-use He4rt\Integrations\Twitch\OAuth\Client\TwitchOAuthClient;
-use He4rt\Integrations\Twitch\OAuth\Contracts\TwitchOAuthService;
-use He4rt\Integrations\Twitch\Subscriber\Client\TwitchSubscribersClient;
-use He4rt\Integrations\Twitch\Subscriber\Contracts\TwitchSubscribersService;
-
-final readonly class TwitchBaseClient implements TwitchService
-{
-    public function __construct(private Client $client) {}
-
-    public function oauth(): TwitchOAuthService
-    {
-        return new TwitchOAuthClient($this->client);
-    }
-
-    public function subscribers(): TwitchSubscribersService
-    {
-        return new TwitchSubscribersClient($this->client);
-    }
-}
