--- app-modules/integrations/src/Twitch/OAuth/Contracts/TwitchOAuthService.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace He4rt\Integrations\Twitch\OAuth\Contracts;
-
-use App\Contracts\OAuthClientContract;
-
-interface TwitchOAuthService extends OAuthClientContract {}
