--- app-modules/integrations/src/Twitch/OAuth/DTO/TwitchOAuthAccessDTO.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace He4rt\Integrations\Twitch\OAuth\DTO;
-
-use He4rt\Identity\Auth\DTOs\OAuthAccessDTO;
-
-class TwitchOAuthAccessDTO extends OAuthAccessDTO
-{
-    public static function make(array $payload): self
-    {
-        return new self(
-            accessToken: $payload['access_token'],
-            refreshToken: $payload['refresh_token'],
-            expiresIn: $payload['expires_in']
-        );
-    }
-}
