--- app-modules/integrations/src/Twitch/Subscriber/Enum/SubscriptionTiersEnum.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace He4rt\Integrations\Twitch\Subscriber\Enum;
-
-enum SubscriptionTiersEnum: string
-{
-    case Tier_1 = '1000';
-    case Tier_2 = '2000';
-    case Tier_3 = '3000';
-}
