+++ composer.json
@@ -24,7 +24,8 @@
         "he4rt/gamification": ">=1",
         "he4rt/he4rt-core": ">=1",
         "he4rt/identity": ">=1",
-        "he4rt/integrations": ">=1",
+        "he4rt/integration-discord": ">=1",
+        "he4rt/integration-twitch": ">=1",
         "he4rt/portal": ">=1",
         "internachi/modular": "dev-main#ad95fe9",
         "laracord/framework": "dev-next",
